lib: create lib.rs and move main logic into it
